fix async issue
