+++ src/stream/src/executor/source.rs
@@ -27,8 +27,8 @@ use risingwave_connector::state::SourceStateHandler;
 use risingwave_connector::{ConnectorState, SplitImpl, SplitMetaData};
 use risingwave_source::*;
 use risingwave_storage::{Keyspace, StateStore};
-use tokio::sync::mpsc::{channel, Receiver, UnboundedReceiver};
-use tokio::sync::{oneshot, Notify};
+use tokio::sync::mpsc::UnboundedReceiver;
+use tokio::sync::{Mutex, Notify};
 use tokio::time::Instant;
 
 use super::error::StreamExecutorError;
@@ -134,7 +134,7 @@ impl<S: StateStore> SourceExecutor<S> {
 
 struct SourceReader {
     /// The reader for stream source.
-    stream_reader: Box<SourceStreamReaderImpl>,
+    stream_reader: Arc<Mutex<Box<SourceStreamReaderImpl>>>,
     /// The reader for barrier.
     barrier_receiver: UnboundedReceiver<Barrier>,
     /// Expected barrier latency in ms. If there are no barrier within the expected barrier
@@ -145,49 +145,34 @@ struct SourceReader {
 impl SourceReader {
     #[try_stream(ok = StreamChunkWithState, error = RwError)]
     async fn stream_reader(
-        mut stream_reader: Box<dyn StreamSourceReader>,
+        stream_reader: Arc<Mutex<Box<SourceStreamReaderImpl>>>,
         notifier: Arc<Notify>,
         expected_barrier_latency_ms: u64,
-        mut inject_source_rx: Receiver<(Box<SourceStreamReaderImpl>, oneshot::Sender<()>)>,
     ) {
-        let (msg_tx, mut msg_rx) = channel::<Result<StreamChunkWithState>>(1);
-        let handler = tokio::task::spawn(async move {
-            loop {
-                let now = Instant::now();
-
-                // We allow data to flow for `expected_barrier_latency_ms` milliseconds.
-                while now.elapsed().as_millis() < expected_barrier_latency_ms as u128 {
-                    tokio::select! {
-                        biased;
-                        reader = inject_source_rx.recv() => {
-                            if let Some((new_reader, tx)) = reader {
-                                stream_reader = new_reader;
-                                tx.send(()).unwrap();
-                            }
-                        }
-                        chunk = stream_reader.next() => { msg_tx.send(chunk).await.unwrap(); }
-                    }
-                }
-
-                // Here we consider two cases:
-                //
-                // 1. Barrier arrived before waiting for notified. In this case, this await will
-                // complete instantly, and we will continue to produce new data.
-                // 2. Barrier arrived after waiting for notified. Then source will be stalled.
-
-                notifier.notified().await;
-            }
-        });
         'outer: loop {
-            match msg_rx.recv().await.unwrap() {
-                Ok(chunk) => yield chunk,
-                Err(e) => {
-                    error!("hang up stream reader due to polling error: {}", e);
-                    handler.abort();
-                    // Then hang up this stream by breaking the loop.
-                    break 'outer;
-                }
+            let now = Instant::now();
+
+            // We allow data to flow for `expected_barrier_latency_ms` milliseconds.
+            while now.elapsed().as_millis() < expected_barrier_latency_ms as u128 {
+                let mut reader_guard = stream_reader.lock().await;
+                let chunk_result = reader_guard.next().await;
+                drop(reader_guard);
+                match chunk_result {
+                    Ok(chunk) => yield chunk,
+                    Err(e) => {
+                        error!("hang up stream reader due to polling error: {}", e);
+                        break 'outer;
+                    }
+                };
             }
+
+            // Here we consider two cases:
+            //
+            // 1. Barrier arrived before waiting for notified. In this case, this await will
+            // complete instantly, and we will continue to produce new data.
+            // 2. Barrier arrived after waiting for notified. Then source will be stalled.
+
+            notifier.notified().await;
         }
 
         futures::future::pending().await
@@ -206,7 +191,6 @@ impl SourceReader {
 
     fn into_stream(
         self,
-        inject_source: Receiver<(Box<SourceStreamReaderImpl>, oneshot::Sender<()>)>,
     ) -> impl Stream<Item = Either<Result<Message>, Result<StreamChunkWithState>>> {
         let notifier = Arc::new(Notify::new());
 
@@ -215,7 +199,6 @@ impl SourceReader {
             self.stream_reader,
             notifier,
             self.expected_barrier_latency_ms,
-            inject_source,
         );
         select_with_strategy(
             barrier_receiver.map(Either::Left),
@@ -326,22 +309,21 @@ impl<S: StateStore> SourceExecutor<S> {
         };
 
         // todo: use epoch from msg to restore state from state store
-        let stream_reader = self
-            .build_stream_source_reader(recover_state)
-            .await
-            .map_err(StreamExecutorError::source_error)?;
+        let stream_reader = Arc::new(Mutex::new(
+            self.build_stream_source_reader(recover_state)
+                .await
+                .map_err(StreamExecutorError::source_error)?,
+        ));
 
         let reader = SourceReader {
-            stream_reader,
+            stream_reader: stream_reader.clone(),
             barrier_receiver,
             expected_barrier_latency_ms: self.expected_barrier_latency_ms,
         };
         yield Message::Barrier(barrier);
 
-        let (inject_source_tx, inject_source_rx) =
-            channel::<(Box<SourceStreamReaderImpl>, oneshot::Sender<()>)>(1);
         #[for_await]
-        for msg in reader.into_stream(inject_source_rx) {
+        for msg in reader.into_stream() {
             match msg {
                 // This branch will be preferred.
                 Either::Left(barrier) => {
@@ -370,20 +352,7 @@ impl<S: StateStore> SourceExecutor<S> {
                                                 ))
                                                 .await
                                                 .map_err(StreamExecutorError::source_error)?;
-
-                                            let (tx, rx) = oneshot::channel();
-                                            inject_source_tx
-                                                .send((reader, tx))
-                                                .await
-                                                .to_rw_result()
-                                                .map_err(|e| {
-                                                    StreamExecutorError::channel_closed(
-                                                        e.to_string(),
-                                                    )
-                                                })?;
-
-                                            // force sync
-                                            rx.await.unwrap();
+                                            *stream_reader.lock().await = reader;
                                             self.stream_source_splits = target_state;
                                         }
                                     }
@@ -473,10 +442,20 @@ mod tests {
     use std::sync::Arc;
 
     use futures::StreamExt;
+    use maplit::hashmap;
     use risingwave_common::array::stream_chunk::StreamChunkTestExt;
     use risingwave_common::array::StreamChunk;
     use risingwave_common::catalog::{ColumnDesc, Field, Schema};
     use risingwave_common::types::DataType;
+    use risingwave_common::util::sort_util::{OrderPair, OrderType};
+    use risingwave_connector::datagen::DatagenSplit;
+    use risingwave_pb::catalog::StreamSourceInfo;
+    use risingwave_pb::data::data_type::TypeName;
+    use risingwave_pb::data::DataType as ProstDataType;
+    use risingwave_pb::plan_common::{
+        ColumnCatalog as ProstColumnCatalog, ColumnDesc as ProstColumnDesc,
+        RowFormatType as ProstRowFormatType,
+    };
     use risingwave_source::*;
     use risingwave_storage::memory::MemoryStateStore;
     use tokio::sync::mpsc::unbounded_channel;
@@ -703,6 +682,173 @@ mod tests {
         executor.next().await.unwrap().unwrap();
         executor.next().await.unwrap().unwrap();
         write_chunk(chunk);
+
+        Ok(())
+    }
+
+    fn mock_stream_source_info() -> StreamSourceInfo {
+        let properties: HashMap<String, String> = hashmap! {
+            "connector".to_string() => "datagen".to_string(),
+            "fields.v1.min".to_string() => "1".to_string(),
+            "fields.v1.max".to_string() => "1000".to_string(),
+            "fields.v1.seed".to_string() => "12345".to_string(),
+            "fields.v2.min".to_string() => "1".to_string(),
+            "fields.v2.max".to_string() => "1000".to_string(),
+            "fields.v2.seed".to_string() => "12345".to_string(),
+        };
+
+        let columns = vec![
+            ProstColumnCatalog {
+                column_desc: Some(ProstColumnDesc {
+                    column_type: Some(ProstDataType {
+                        type_name: TypeName::Int64 as i32,
+                        ..Default::default()
+                    }),
+                    column_id: 0,
+                    ..Default::default()
+                }),
+                is_hidden: false,
+            },
+            ProstColumnCatalog {
+                column_desc: Some(ProstColumnDesc {
+                    column_type: Some(ProstDataType {
+                        type_name: TypeName::Int32 as i32,
+                        ..Default::default()
+                    }),
+                    column_id: 1,
+                    name: "v1".to_string(),
+                    ..Default::default()
+                }),
+                is_hidden: false,
+            },
+            ProstColumnCatalog {
+                column_desc: Some(ProstColumnDesc {
+                    column_type: Some(ProstDataType {
+                        type_name: TypeName::Float as i32,
+                        ..Default::default()
+                    }),
+                    column_id: 2,
+                    name: "v2".to_string(),
+                    ..Default::default()
+                }),
+                is_hidden: false,
+            },
+        ];
+
+        StreamSourceInfo {
+            properties,
+            row_format: ProstRowFormatType::Json as i32,
+            row_schema_location: "".to_string(),
+            row_id_index: 0,
+            columns,
+            pk_column_ids: vec![0],
+        }
+    }
+
+    #[tokio::test]
+    async fn test_split_change_mutation() -> Result<()> {
+        let stream_source_info = mock_stream_source_info();
+        let source_table_id = TableId::default();
+        let source_manager = Arc::new(MemSourceManager::default());
+
+        source_manager
+            .create_source(&source_table_id, stream_source_info)
+            .await?;
+
+        let get_schema = |column_ids: &[ColumnId], source_desc: &SourceDesc| {
+            let mut fields = Vec::with_capacity(column_ids.len());
+            for &column_id in column_ids {
+                let column_desc = source_desc
+                    .columns
+                    .iter()
+                    .find(|c| c.column_id == column_id)
+                    .unwrap();
+                fields.push(Field::unnamed(column_desc.data_type.clone()));
+            }
+            Schema::new(fields)
+        };
+
+        let actor_id = ActorId::default();
+        let source_desc = source_manager.get_source(&source_table_id)?;
+        let keyspace = Keyspace::table_root(MemoryStateStore::new(), &TableId::from(0x2333));
+        let column_ids = vec![ColumnId::from(0), ColumnId::from(1), ColumnId::from(2)];
+        let schema = get_schema(&column_ids, &source_desc);
+        let pk_indices = vec![0_usize];
+        let (barrier_tx, barrier_rx) = unbounded_channel::<Barrier>();
+
+        let source_exec = SourceExecutor::new(
+            actor_id,
+            source_table_id,
+            source_desc,
+            keyspace.clone(),
+            column_ids.clone(),
+            schema,
+            pk_indices,
+            barrier_rx,
+            1,
+            1,
+            "SourceExecutor".to_string(),
+            Arc::new(StreamingMetrics::unused()),
+            u64::MAX,
+        )?;
+
+        let mut materialize = MaterializeExecutor::new(
+            Box::new(source_exec),
+            keyspace.clone(),
+            vec![OrderPair::new(0, OrderType::Ascending)],
+            column_ids.clone(),
+            2,
+            vec![0usize],
+        )
+        .boxed()
+        .execute();
+
+        let curr_epoch = 1919;
+        let init_barrier =
+            Barrier::new_test_barrier(curr_epoch).with_mutation(Mutation::AddOutput(AddOutput {
+                map: HashMap::new(),
+                splits: hashmap! {
+                    ActorId::default() => vec![
+                        SplitImpl::Datagen(
+                        DatagenSplit {
+                            split_index: 0,
+                            split_num: 3,
+                            start_offset: None,
+                        }),
+                    ],
+                },
+            }));
+        barrier_tx.send(init_barrier).unwrap();
+
+        let _ = materialize.next().await.unwrap(); // barrier
+
+        let _ = materialize.next().await.unwrap();
+
+        let change_split_mutation = Barrier::new_test_barrier(curr_epoch + 1).with_mutation(
+            Mutation::SourceChangeSplit(hashmap! {
+                ActorId::default() => Some(vec![
+                    SplitImpl::Datagen(
+                        DatagenSplit {
+                            split_index: 0,
+                            split_num: 3,
+                            start_offset: None,
+                        }
+                    ), SplitImpl::Datagen(
+                        DatagenSplit {
+                            split_index: 1,
+                            split_num: 3,
+                            start_offset: None,
+                        }
+                    ),
+                ])
+            }),
+        );
+        barrier_tx.send(change_split_mutation).unwrap();
+
+        let _ = materialize.next().await.unwrap(); // barrier
+
+        let _ = materialize.next().await.unwrap();
+        let _ = materialize.next().await.unwrap();
 
         Ok(())
     }
